Remove the manual installation section from the self-hosting page
The manual section suggests using the docker-compose.yml and .env from
our repo, but those are development files, requires the source code, and
should not be used by someone trying to setup Appwrite.

We need to figure out how to automate the creation of the
docker-compose.yml and .env files and then bring back this section.