fix: remove envFixer plugin to remove zero-width spaces
The plugin was created because `process.env` in our code samples was
being replaced so the code samples looked incorrect. However, Vite 5
has fixed this so we don't need this plugin anymore.

In addition, the zero-width space was causing issues when people
would copy and paste the code into their editor to run.