Added feature that the logged in user is displayed with logout button instead after reloading
As per the code in docs, after logging in, if we refresh the page then the loggedInUser state is set on null as default and it will be changed only when the login function is called, but the user is still logged in. so i have integrated an feature that if the user is logged in then the logged in component will be displayed.