feat(init): update /init/tickets/update to lookup the ticket id
Since there's a cookie with the session secret, we can use that to make
an authenticated request to Appwrite to get details about the user.

Since we lookup the ticket document using the GitHub username, we need
to fetch the username from GitHub using the providerAccessToken.

To get the providerAccessToken, we need to use account.listIdentities()
because there's a bug where the providerAccessToken is not available in
the session.