fix(tutorial): update redirect in SvelteKit SSR auth tutorial
The current redirect status code is 301, but that is a permanent
redirect. The problem with this is the browser will cache the redirect
and redirect for every request without even making a request to the
server.

Instead, we should use a 302 status code, which is a temporary redirect.
So, when the user hits the home page, the browser will hit the server
so that the server checks whether the user is logged in or not and
redirect accordingly instead of always redirecting to the same page.