feat: decouple support from billing plan
There are orgs who are on lower billing plans, but want higher support levels.

There are orgs who are on higher billing plans, but don't care about support.

This change allows us to decouple the two so that customers can get the right support level for their needs, regardless of their billing plan.